Move the forms for the admissions in a new module
